fix kerning from 7dc518624c5bd10b4ada17f1b19ac0292d731e3c merge

+++ lib/EpdFont/EpdFont.cpp
@@ -95,7 +95,7 @@ static uint8_t lookupKernClass(const EpdKernClassEntry* entries, const uint16_t 
   return 0;
 }
 
-int8_t EpdFont::getKerning(const uint32_t leftCp, const uint32_t rightCp) const {
+int EpdFont::getKerning(const uint32_t leftCp, const uint32_t rightCp) const {
   if (!data->kernMatrix) {
     return 0;
   }
