+++ lib/EpdFont/EpdFont.h
@@ -14,7 +14,7 @@ class EpdFont {
 
   /// Returns the kerning adjustment (4.4 fixed-point in pixels) between two codepoints.
   /// Returns 0 if no kerning data exists for the pair.
-  int8_t getKerning(uint32_t leftCp, uint32_t rightCp) const;
+  int getKerning(uint32_t leftCp, uint32_t rightCp) const;
 
   /// Returns the ligature codepoint for a pair, or 0 if no ligature exists.
   uint32_t getLigature(uint32_t leftCp, uint32_t rightCp) const;
